LPS-201369 move out of date banner to toolbar so it shows for all views
