+++ modules/apps/change-tracking/change-tracking-web/src/main/resources/META-INF/resources/publications/js/views/ChangeTrackingChangesToolbar.js
@@ -3,10 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
+import ClayAlert from '@clayui/alert';
 import ClayButton, {ClayButtonWithIcon} from '@clayui/button';
 import {ClayDropDownWithItems} from '@clayui/drop-down';
 import ClayIcon from '@clayui/icon';
 import ClayLabel from '@clayui/label';
+import ClayLayout from '@clayui/layout';
 import ClayToolbar from '@clayui/toolbar';
 import classNames from 'classnames';
 import React, {useCallback, useRef, useState} from 'react';
@@ -192,6 +194,24 @@ export default function ChangeTrackingChangesToolbar({
 		);
 	};
 
+	const renderExpiredBanner = () => {
+		if (!expired) {
+			return '';
+		}
+
+		return (
+			<ClayAlert
+				displayType="warning"
+				spritemap={spritemap}
+				title={Liferay.Language.get('out-of-date')}
+			>
+				{Liferay.Language.get(
+					'this-publication-was-created-on-a-previous-liferay-version.-you-cannot-publish,-revert,-or-make-additional-changes'
+				)}
+			</ClayAlert>
+		);
+	};
+
 	return (
 		<>
 			{renderPublicationsToolbar()}
@@ -252,6 +272,10 @@ export default function ChangeTrackingChangesToolbar({
 					</div>
 				</div>
 			</div>
+
+			<ClayLayout.ContainerFluid style={{marginTop: '1em'}}>
+				{renderExpiredBanner()}
+			</ClayLayout.ContainerFluid>
 		</>
 	);
 }
